Added priority, fixed json issues, etc (wip)

+++ apps/backend/src/main/java/no/nav/data/integration/slack/SlackService.java
@@ -3,7 +3,7 @@ package no.nav.data.integration.slack;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.javacrumbs.shedlock.spring.annotation.SchedulerLock;
-import no.nav.data.integration.slack.dto.SlackDtos.PostMessageRequest.Block;
+import no.nav.data.integration.slack.dto.SlackDtos.Block;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
@@ -20,38 +20,50 @@ public class SlackService {
     private final SlackClient slackClient;
 
     @Transactional(propagation = Propagation.REQUIRED)
-    public void sendMessageToChannel(String mottager, List<Block> blocks) {
-        SlackMelding sm = new SlackMelding(mottager, false, blocks);
+    public void sendMessageToChannel(String mottager, List<Block> blocks, int priority) {
+        SlackMelding sm = new SlackMelding(mottager, false, priority, blocks);
         repo.save(sm);
     }
 
     @Transactional(propagation = Propagation.REQUIRED)
-    public void sendMessageToUser(String mottager, List<Block> blocks) {
-        SlackMelding sm = new SlackMelding(mottager, true, blocks);
+    public void sendMessageToUser(String mottager, List<Block> blocks, int priority) {
+        SlackMelding sm = new SlackMelding(mottager, true, priority, blocks);
         repo.save(sm);
     }
     
-    @SchedulerLock(name = "sendSlack")
-    @Scheduled(cron = "0 0 13 * * *") // Happens every day at 1300
+    @SchedulerLock(name = "sendSlackEnGros")
+    @Scheduled(cron = "0 55 12 * * *") // Happens every day at 12:55:00
     public void sendAll() {
         log.info("Sending all pending slack messages...");
         int sendCount = 0;
-        while (repo.count() > 0) {
-            sendOneMessage();
+        while (sendOneMessage(SlackMelding.PRIORITY_LOW)) {
             sendCount++;
         }
         log.info("Done sending {} pending slack messages", sendCount);
     }
 
+    @SchedulerLock(name = "sendSlackPriority")
+    @Scheduled(cron = "10 * * * * *") // Happens every minute at 10 seconds.
+    public void sendHighPriority() {
+        int sendCount = 0;
+        while (sendOneMessage(SlackMelding.PRIORITY_HIGH)) {
+            sendCount++;
+        }
+        log.info("Done sending {} high priority slack messages", sendCount);
+    }
+
     @Transactional()
-    private void sendOneMessage() {
-        SlackMelding sm = repo.getOne();
+    // Returns true if there was a pending message to send, false otherwise.
+    private boolean sendOneMessage(int priority) {
+        SlackMelding sm = repo.getOneWithPriority(priority).orElse(null);
+        if (sm == null) return false;
         if (sm.getSendTilKanal()) {
             slackClient.sendMessageToChannel(sm.getMottager(), sm.getBlocks());
         } else {
             slackClient.sendMessageToUserId(sm.getMottager(), sm.getBlocks());
         }
         repo.delete(sm);
+        return true;
     }
     
 }
